pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "comic-text-separator"
version = "0.2.0"
description = "Extract and separate text layers from comic/conti images to PSD or JSON"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "GPL-3.0"}
authors = [{name = "suwonleee"}]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Environment :: GPU :: NVIDIA CUDA",
]

# PyPI serves CPU-only torch on Mac/Windows by default.
# Install scripts override with GPU variants when appropriate.
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "einops",
    "numpy",
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
    "shapely",
    "pyclipper",
    "networkx",
    "py3langid",
    "colorama",
    "tqdm",
    "requests",
    "psd-tools>=1.12",
    "kss>=6.0",
    "PySide6>=6.6",
]

[project.scripts]
comic-sep = "main:main"

[project.gui-scripts]
comic-sep-gui = "gui.app:gui_main"

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.1.0"]

[tool.setuptools.packages.find]
include = ["domain*", "use_cases*", "adapters*", "engine*", "infrastructure*", "gui*"]

[tool.setuptools]
py-modules = ["main"]
